// src/models/IERC20.ts
export interface IERC20Interface extends Interface {
	getFunction(
		nameOrSignature:
			| "allowance"
			| "approve"
			| "balanceOf"
			| "decimals"
			| "name"
			| "symbol"
			| "totalSupply"
			| "transfer"
			| "transferFrom"
	): FunctionFragment;

	getEvent(nameOrSignatureOrTopic: "Approval" | "Transfer"): EventFragment;

	encodeFunctionData(
		functionFragment: "allowance",
		values: [AddressLike, AddressLike]
	): string;
	encodeFunctionData(
		functionFragment: "approve",
		values: [AddressLike, BigNumberish]
	): string;
	encodeFunctionData(
		functionFragment: "balanceOf",
		values: [AddressLike]
	): string;
	encodeFunctionData(functionFragment: "decimals", values?: undefined): string;
	encodeFunctionData(functionFragment: "name", values?: undefined): string;
	encodeFunctionData(functionFragment: "symbol", values?: undefined): string;
	encodeFunctionData(
		functionFragment: "totalSupply",
		values?: undefined
	): string;
	encodeFunctionData(
		functionFragment: "transfer",
		values: [AddressLike, BigNumberish]
	): string;
	encodeFunctionData(
		functionFragment: "transferFrom",
		values: [AddressLike, AddressLike, BigNumberish]
	): string;

	decodeFunctionResult(functionFragment: "allowance", data: BytesLike): Result;
	decodeFunctionResult(functionFragment: "approve", data: BytesLike): Result;
	decodeFunctionResult(functionFragment: "balanceOf", data: BytesLike): Result;
	decodeFunctionResult(functionFragment: "decimals", data: BytesLike): Result;
	decodeFunctionResult(functionFragment: "name", data: BytesLike): Result;
	decodeFunctionResult(functionFragment: "symbol", data: BytesLike): Result;
	decodeFunctionResult(
		functionFragment: "totalSupply",
		data: BytesLike
	): Result;
	decodeFunctionResult(functionFragment: "transfer", data: BytesLike): Result;
	decodeFunctionResult(
		functionFragment: "transferFrom",
		data: BytesLike
	): Result;
}

export namespace ApprovalEvent {
	export type InputTuple = [
		owner: AddressLike,
		spender: AddressLike,
		value: BigNumberish
	];
	export type OutputTuple = [owner: string, spender: string, value: bigint];
	export interface OutputObject {
		owner: string;
		spender: string;
		value: bigint;
	}
	export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
	export type Filter = TypedDeferredTopicFilter<Event>;
	export type Log = TypedEventLog<Event>;
	export type LogDescription = TypedLogDescription<Event>;
}

export namespace TransferEvent {
	export type InputTuple = [
		from: AddressLike,
		to: AddressLike,
		value: BigNumberish
	];
	export type OutputTuple = [from: string, to: string, value: bigint];
	export interface OutputObject {
		from: string;
		to: string;
		value: bigint;
	}
	export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
	export type Filter = TypedDeferredTopicFilter<Event>;
	export type Log = TypedEventLog<Event>;
	export type LogDescription = TypedLogDescription<Event>;
}

export interface IERC20 extends BaseContract {
	connect(runner?: ContractRunner | null): IERC20;
	waitForDeployment(): Promise<this>;

	interface: IERC20Interface;

	queryFilter<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		fromBlockOrBlockhash?: string | number | undefined,
		toBlock?: string | number | undefined
	): Promise<Array<TypedEventLog<TCEvent>>>;
	queryFilter<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		fromBlockOrBlockhash?: string | number | undefined,
		toBlock?: string | number | undefined
	): Promise<Array<TypedEventLog<TCEvent>>>;

	on<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		listener: TypedListener<TCEvent>
	): Promise<this>;
	on<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		listener: TypedListener<TCEvent>
	): Promise<this>;

	once<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		listener: TypedListener<TCEvent>
	): Promise<this>;
	once<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		listener: TypedListener<TCEvent>
	): Promise<this>;

	listeners<TCEvent extends TypedContractEvent>(
		event: TCEvent
	): Promise<Array<TypedListener<TCEvent>>>;
	listeners(eventName?: string): Promise<Array<Listener>>;
	removeAllListeners<TCEvent extends TypedContractEvent>(
		event?: TCEvent
	): Promise<this>;

	allowance: TypedContractMethod<
		[owner: AddressLike, spender: AddressLike],
		[bigint],
		"view"
	>;

	approve: TypedContractMethod<
		[spender: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;

	balanceOf: TypedContractMethod<[account: AddressLike], [bigint], "view">;

	decimals: TypedContractMethod<[], [bigint], "view">;

	name: TypedContractMethod<[], [string], "view">;

	symbol: TypedContractMethod<[], [string], "view">;

	totalSupply: TypedContractMethod<[], [bigint], "view">;

	transfer: TypedContractMethod<
		[to: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;

	transferFrom: TypedContractMethod<
		[from: AddressLike, to: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;

	getFunction<T extends ContractMethod = ContractMethod>(
		key: string | FunctionFragment
	): T;

	getFunction(
		nameOrSignature: "allowance"
	): TypedContractMethod<
		[owner: AddressLike, spender: AddressLike],
		[bigint],
		"view"
	>;
	getFunction(
		nameOrSignature: "approve"
	): TypedContractMethod<
		[spender: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;
	getFunction(
		nameOrSignature: "balanceOf"
	): TypedContractMethod<[account: AddressLike], [bigint], "view">;
	getFunction(
		nameOrSignature: "decimals"
	): TypedContractMethod<[], [bigint], "view">;
	getFunction(
		nameOrSignature: "name"
	): TypedContractMethod<[], [string], "view">;
	getFunction(
		nameOrSignature: "symbol"
	): TypedContractMethod<[], [string], "view">;
	getFunction(
		nameOrSignature: "totalSupply"
	): TypedContractMethod<[], [bigint], "view">;
	getFunction(
		nameOrSignature: "transfer"
	): TypedContractMethod<
		[to: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;
	getFunction(
		nameOrSignature: "transferFrom"
	): TypedContractMethod<
		[from: AddressLike, to: AddressLike, value: BigNumberish],
		[boolean],
		"nonpayable"
	>;

	getEvent(
		key: "Approval"
	): TypedContractEvent<
		ApprovalEvent.InputTuple,
		ApprovalEvent.OutputTuple,
		ApprovalEvent.OutputObject
	>;
	getEvent(
		key: "Transfer"
	): TypedContractEvent<
		TransferEvent.InputTuple,
		TransferEvent.OutputTuple,
		TransferEvent.OutputObject
	>;

	filters: {
		"Approval(address,address,uint256)": TypedContractEvent<
			ApprovalEvent.InputTuple,
			ApprovalEvent.OutputTuple,
			ApprovalEvent.OutputObject
		>;
		Approval: TypedContractEvent<
			ApprovalEvent.InputTuple,
			ApprovalEvent.OutputTuple,
			ApprovalEvent.OutputObject
		>;

		"Transfer(address,address,uint256)": TypedContractEvent<
			TransferEvent.InputTuple,
			TransferEvent.OutputTuple,
			TransferEvent.OutputObject
		>;
		Transfer: TypedContractEvent<
			TransferEvent.InputTuple,
			TransferEvent.OutputTuple,
			TransferEvent.OutputObject
		>;
	};
}